backend/requirements.txt
Django>=4.2.0,<5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.5
setuptools>=65.5.1